feat: work on avoiding circular dependencies
